Move extra method.
